Module arcs::algorithms

Common algorithms.

Structs

ApproximatedArc
An iterator over the points in an arc approximation.

Enums

Closest
An enum containing the different possible solutions for ClosestPoint::closest_point().

Traits

AffineTransformable
Something which can be transformed using an arbitrary Transform2D matrix and still be semantically valid.

Approximate
Approximate a shape with a bunch of Points.

Bounded
Calculate an axis-aligned bounding box around the item.

ClosestPoint
Find the location on an object which is closest to a target point.

Length
Something which has a finite length.

Scale
Something who's dimensions can be scaled uniformly.

ScaleNonUniform
Something who's dimensions can be scaled independently (the non-uniform bit) in the x and y directions.

Translate
Something which can be moved around "rigidly" in Drawing Space.

Functions

simplify
Decimate a curve composed of line segments to a "simpler" curve with fewer points.