Module arcs::components

Common components used by the arcs CAD library.

Structs

BoundingBox
An axis-aligned bounding box.

DrawingObject
Something which can be drawn on the screen.

Layer
A logical grouping of data, assembled as though each Layer were laid out on transparent acetate overlays.

LineStyle

Name
A name that can be looked up later in the NameTable.

NameTable
A global Resource for looking up an Entity using its Name.

PointStyle

Selected
An empty Component used to mark an Entity as selected.

Space
A global [Resource] for looking up which Entitys inhabit a given spatial point or region

SpatialEntity
A intermediate struct that maps an Entity to its BoundingBox

Viewport

WindowStyle

Enums

Dimension
A dimension on the canvas.

Geometry
The geometry of a DrawingObject.

Functions

register
Register all specs::Components.