Struct arcs::primitives::Arc

pub struct Arc<S> { /* fields omitted */ }

A circle segment.

Implementations

impl<S> Arc<S>

pub fn from_centre_radius(
centre: Point2D<f64, S>,
radius: f64,
start_angle: Angle<f64>,
sweep_angle: Angle<f64>
) -> Arc<S>

Create an Arc based upon its centre and radius.

Examples

pub fn from_three_points(
start: Point2D<f64, S>,
middle: Point2D<f64, S>,
end: Point2D<f64, S>
) -> Option<Arc<S>>

Try to find the Arc which will pass through three points.

Examples

You can use this constructor in the normal way.

use arcs_core::primitives::Arc;

let right = Point::new(10.0, 0.0);
let above = Point::new(0.0, 10.0);
let left = Point::new(-10.0, 0.0);

let got = Arc::from_three_points(right, above, left).unwrap();

assert_eq!(got.centre(), Point::zero());
assert_eq!(got.radius(), 10.0);
assert!(got.is_anticlockwise());

This will fail if the three points are Orientation::Collinear.

use arcs_core::primitives::Arc;

let start = Point::new(0.0, 0.0);
let middle = Point::new(10.0, 0.0);
let end = Point::new(20.0, 0.0);

let got = Arc::from_three_points(start, middle, end);

assert!(got.is_none());

pub const fn centre(self) -> Point2D<f64, S>

The Arc's centre point.

pub const fn radius(self) -> f64

The Arc's radius.

pub const fn start_angle(self) -> Angle<f64>

pub const fn sweep_angle(self) -> Angle<f64>

pub fn end_angle(self) -> Angle<f64>

pub fn is_anticlockwise(self) -> bool

pub fn is_clockwise(self) -> bool

pub fn start(self) -> Point2D<f64, S>

pub fn end(self) -> Point2D<f64, S>

pub fn point_at(self, angle: Angle<f64>) -> Point2D<f64, S>

pub fn contains_angle(self, angle: Angle<f64>) -> bool

pub fn is_minor_arc(&self) -> bool

pub fn is_major_arc(&self) -> bool